Major overhaul to the physics system
PhysicsBody now checks on the relative speeds of
the objects and the direction relative to the collision
surface before triggering collision effects.

Added an experimental but mostly unfinished torque
calculation system that will impart momentum on collision.
(Has to be fixed)

Fixed bug in PhysicsManager where the impulse direction/
normal of the colliding surface was added with the wrong direction
to object a if object A had reflectionOnCollision set to true.